:bug: made time 64 bit

+++ include/transactions.hpp
@@ -2,13 +2,13 @@
 #include <chrono>
 #include <cstdint>
 
-//24 bytes total
+//42 bytes total
 struct Transaction
 {
 
     std::string from = "", to = "";
     uint32_t amount = 0;
-    uint32_t time = 0;
+    uint64_t time = 0;
 
     void Concatinate(std::string &s)
     {
